requirements.txt
arabic-reshaper==3.0.0
bidict==0.23.1
blinker==1.9.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
greenlet==3.2.3
h11==0.16.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
PyPDF2==3.0.1
python-bidi==0.6.6
python-dotenv==0.19.0
python-engineio==4.12.2
python-socketio==5.13.0
reportlab==4.4.2
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
webencodings==0.5.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
xhtml2pdf==0.2.5
